refactor: gather example tests in single test file, don't use importorskip
