--- tests/add_test.py
@@ -1,5 +1,0 @@
-from pytest import importorskip
-
-
-def test_add():
-    importorskip("tests.assets.add")
