--- tests/hello_test.py
@@ -1,5 +1,0 @@
-from pytest import importorskip
-
-
-def test_hello():
-    importorskip("tests.assets.hello")
