--- tests/rich_versions_test.py
@@ -1,5 +1,0 @@
-from pytest import importorskip
-
-
-def test_rich_versions():
-    importorskip("tests.assets.rich_versions")
