--- tests/simple_test.py
@@ -1,5 +1,0 @@
-from pytest import importorskip
-
-
-def test_simple():
-    importorskip("tests.assets.simple")
